ServiceDeskBot-1404/backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
requests
sqlalchemy
pydantic
psycopg2-binary
python-dotenv